Fixes item verbs
Hopefully this helps. You were returning a default value in the update function too early, so it wasn't getting ot the get, inspect, or action/item parts of that function. Also, in items, the verb needs to be capitalized (sorry if that was a typo in the file I gave you).